fix for worker-loader
