--- typings/worker-loader.d.ts
@@ -1,10 +1,0 @@
-declare module "worker-loader!*" {
-    // You need to change `Worker`, if you specified a different value for the `workerType` option
-    class WebpackWorker extends Worker {
-      constructor();
-    }
-  
-    // Uncomment this if you set the `esModule` option to `false`
-    // export = WebpackWorker;
-    export default WebpackWorker;
-  }
